SDNQ skip bad layers on svd and fix svd with dequantize_fp32

+++ modules/sdnq/layers/conv/conv_fp8_tensorwise.py
@@ -29,7 +29,7 @@ def conv_fp8_matmul_tensorwise(
     if svd_up is not None:
         input = input.flatten(0,-2)
         if bias is not None:
-            bias = torch.addmm(bias, torch.mm(input.to(dtype=svd_down.dtype), svd_down), svd_up)
+            bias = torch.addmm(bias.to(dtype=svd_down.dtype), torch.mm(input.to(dtype=svd_down.dtype), svd_down), svd_up)
         else:
             bias = torch.mm(torch.mm(input.to(dtype=svd_down.dtype), svd_down), svd_up)
